Implement max side length (north)
This is for
https://leetcode.com/problems/get-biggest-three-rhombus-sums-in-a-grid/.

I want to figure out how far out I can go when tracing a rhombus.

If I'm tracing a rhombus on an 8x8 from (1, 3), I want to head down to
(4, 0) and start building up memos of the lengths. I don't want to start
memo-ing from the northern corner, because I'll lever encounter sides
including that square again.